implemented next_avavilable_ip function both dynamic and static
