+++ example/go.mod
@@ -2,6 +2,6 @@ module github.com/dupparagonabhavi-infoblox/example
 
 go 1.24.1
 
-require github.com/dupparagonabhavi-infoblox/infoblox_go_sdk/output/go-sdk v0.0.0-20250410145020-6aac43bf6ffa
+require github.com/dupparagonabhavi-infoblox/infoblox_go_sdk/output/go-sdk v0.0.0-20250411172007-f7f24ce4b971
 
 require gopkg.in/validator.v2 v2.0.1 // indirect
